Lots of bugfixes, added some logic to differentiate between updating and initial setup.

+++ setup.ps1
@@ -24,7 +24,6 @@ $npmGlobalPackages = @(
 	"typescript"
 	"browser-sync"
 )
-
 $npmSaveDevPackages = @(
 	"gulp"
 	"del"
@@ -60,6 +59,21 @@ $npmSaveDevPackages = @(
 #	"gulp-pug"
 )
 
+$NPMPath = "C:\Program Files\nodejs"
+$GITPath = "C:\Program Files\Git\cmd"
+
+#Determine if updating or setting up for the first time.
+{
+	$scriptname = $MyInvocation.MyCommand.Name;
+	if($scriptname -eq "setup.ps1") { $SettingUp = "true" }
+	if($scriptname -eq "update.ps1") { $Updating = "true" }
+	if(($SettingUp -ne "true") -and ($Updating -ne "true"))
+	{
+		Write-Output "Unable to detirmine if updating or setting up. Please make sure this powershell script is named update.ps1 or setup.ps1"
+		return
+	}
+}
+
 Write-Output "Checking to see if Chocolatey is installed..."
 
 if (Get-Command choco -errorAction SilentlyContinue) {
@@ -71,35 +85,48 @@ if (Get-Command choco -errorAction SilentlyContinue) {
 }
 else {
 	Write-Output "Chocolatey isn't installed, attempting to install. "
-	Set-ExecutionPolicy Bypass -Scope Process -Force; Invok-Expression ((New-Object System.Net.WebClient).DownloadString('https://chocolatey.org/install.ps1'))
+	Set-ExecutionPolicy Bypass -Scope Process -Force; Invoke-Expression ((New-Object System.Net.WebClient).DownloadString('https://chocolatey.org/install.ps1'))
 
 	Write-Output "Attempting to install apps with chocolatey"
 	choco install $Applications -y
 }
 
-Write-Output "Checking to see if npm is installed..."
+if(-not ((Get-Command npm -errorAction SilentlyContinue))) {
+	$Env:Path = "$Env:Path;$NPMPath"
+}
 
 if (Get-Command npm -errorAction SilentlyContinue) {
 	Write-Output "Installing NPM packages..."
 
 #	PreOP
 	npm --silent install  npm@latest -g
-	npm --silent cache clean -f
 	npm --silent install node@latest -g --scripts-prepend-node-path
-	npm --silent init -f
-	npm --silent rm --global gulp
+	if ($SettingUp -eq "true")
+	{
+		npm --silent init -f
+		npm --silent rm --global gulp
+		npm --silent cache clean -f
+	}
 
-#	Actually install packages
+#	install/update packages
 	npm --silent install -g $npmGlobalPackages
 	npm --silent install -save-dev $npmSaveDevPackages
 
 #	PostOP
-	tslint --init
-	npm --silent audit fix
-	remove-item package.json
+	npm --silent audit fix	
+	
+	if($SettingUp -eq "true")
+	{
+		tslint --init
+	}
 }
 else {
 	Write-Output "NPM command wasn't found, there may have been a problem during installation, try running this script again."
+	return
+}
+
+if(-not ((Get-Command git -errorAction SilentlyContinue))) {
+	$Env:Path = "$Env:Path;$GITPath"
 }
 
 if (Get-Command git -errorAction SilentlyContinue) {
@@ -112,4 +139,11 @@ if (Get-Command git -errorAction SilentlyContinue) {
 }
 else {
 	Write-Output "GIT command wasn't found, there may have been a problem during installation, try running this script again."
+	return
+}
+
+if($SettingUp -eq "true")
+{
+	Set-Location ..
+	try { rename-item "setup.ps1" "update.ps1" } catch {}
 }
